backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.28.1
pydub==0.25.1
librosa==0.10.2
midiutil==1.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
numpy==2.1.0
soundfile==0.12.1
pedalboard==0.9.22
